pyproject.toml
[project]
name = "goobot"
version = "0.1.0"
description = "Goobot = goo discord bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "pydantic>=2.11.9",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
